feat(control_editor): add basic editing, event handling
- Implement basic control information editing, including position and size (absolute and percentage).
- Add click event editing for controls, allowing assignment of launcher-specific actions and key presses.
- Introduce style editing capabilities (placeholder for now).
- Enhance the control editor UI with a tabbed layout for better organization of editing options.
- Allow users to preview changes in real-time by making the editing dialog semi-transparent.
- Refactor control editor logic into a dedicated ViewModel.
- Improve slider components for better usability and visual feedback.

BREAKING CHANGE: The `ButtonPosition` data class now uses `Int` for x and y coordinates (0-1000 range) instead of `Offset`. `ButtonSize` percentage values are now `Int` (0-1000 range) instead of `Float`.

+++ LayerController/src/main/java/com/movtery/layer_controller/data/ButtonSize.kt
@@ -10,9 +10,9 @@ data class ButtonSize(
     @SerializedName("heightDp")
     val heightDp: Float,
     @SerializedName("widthPercentage")
-    val widthPercentage: Float,
+    val widthPercentage: Int,
     @SerializedName("heightPercentage")
-    val heightPercentage: Float,
+    val heightPercentage: Int,
     @SerializedName("widthReference")
     val widthReference: Reference,
     @SerializedName("heightReference")
@@ -58,8 +58,8 @@ data class ButtonSize(
             type = Type.Percentage,
             widthDp = 50f,
             heightDp = 50f,
-            widthPercentage = 140f,
-            heightPercentage = 140f,
+            widthPercentage = 140,
+            heightPercentage = 140,
             widthReference = Reference.ScreenHeight,
             heightReference = Reference.ScreenHeight
         )
